check valid_p_hook_removed_diagrams is empty list
if you only check if the list of p_hooks is empty, you miss the case when there are valid p-hooks, but removing them doesn't result in a valid diagram. the check should be performed right before the recursive step. it's saying:

if you can remove a p-hook and get a valid diagram, keep going
if you can't, you're done