note exactly the difference between Schur's matrix and the DFT
one needs to apply the permutation \sigma = (1 2 ... n) on rows and columns to get from Schur's matrix to the DFT.

let P_\sigma be the row permutation. then P_\sigma^T = P_\sigma^{-1} is the column transformation. the eigenvalues will the same but the eigenvectors will be Pv, so a cyclic shift.