find eigenvalues in splitting field
we can form the splitting field L = K[x]/char_poly(x) directly via L.<alpha> = K.extension(char_poly(x)). for n=3 this gives a polynomial of degree 6 which we can view over L. Then this splits, but it is impossible to express the roots in terms base square roots - you get \alpha and a quintic. It appears one needs Galois theory at this point, so the other roots are Galois conjugates of \alpha.