solution as matrix
format the solution vector as a d_rho x d_rho matrix. this is the sesquilinear form we're looking for. now need to verify it is sesquilinear, and check for which partitions we get solutions.

for cyclic group case, there is no bug, it's just for some values of q, n you don't get a unitary matrix.